+++ shapely-core/src/impls/mod.rs
