+++ shapely-core/src/lib.rs
@@ -1,7 +1,11 @@
 #![warn(missing_docs)]
 #![doc = include_str!("../README.md")]
 
-use std::mem::MaybeUninit;
+mod opaque;
+pub use opaque::*;
+
+mod vtable;
+pub use vtable::*;
 
 mod impls;
