+++ shapely-core/src/opaque.rs
