+++ shapely-core/src/peek/mod.rs
