+++ shapely-core/src/poke/list.rs
@@ -1,0 +1,8 @@
+use crate::{ListVTable, Opaque, Shape, ValueVTable};
+
+pub struct PokeList<'mem> {
+    pub data: Opaque<'mem>,
+    pub shape: Shape,
+    pub vtable: ValueVTable,
+    pub list_vtable: ListVTable,
+}
