+++ shapely-core/src/vtable.rs
@@ -1,0 +1,14 @@
+mod value;
+pub use value::*;
+
+mod struct_;
+pub use struct_::*;
+
+mod enum_;
+pub use enum_::*;
+
+mod list;
+pub use list::*;
+
+mod map;
+pub use map::*;
