+++ shapely-core/src/vtable/enum_.rs
