+++ shapely-core/src/vtable/list.rs
