+++ shapely-core/src/vtable/map.rs
