+++ shapely-core/src/vtable/struct_.rs
